requirements.txt
# Article Scraper Dependencies
# Created: 2025-10-09

# HTTP Server (for future HTTP mode)
flask==3.0.3

# Database access (for poller mode)
psycopg2-binary==2.9.10

# Primary article extraction
trafilatura==1.12.2

# Fallback extraction (news-specific)
newspaper3k==0.2.8

# HTTP client (async support)
aiohttp==3.10.10
requests==2.32.3

# HTML parsing (used by newspaper3k)
lxml==5.3.0
beautifulsoup4==4.12.3

# Optional: Better date parsing
python-dateutil==2.9.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
